requirements.txt
fastapi
sqlalchemy
databases
asyncpg
psycopg2
uvicorn
passlib[bcrypt]
python-jose
